backend-python/requirements.txt
mysql-connector-python
google-auth